requirements.txt
requests==2.32.3
PyMuPDF==1.24.9
ocrmypdf==16.4.3
Pillow==10.4.0
markdownify==0.12.1
PyYAML==6.0.2
google-cloud-aiplatform>=1.68.0,<2.0.0
google-auth>=2.20.0
google-api-core>=2.10.0

